requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery[redis]==5.4.0
redis==5.0.4
httpx==0.27.0
openai==1.30.1
sentence-transformers==3.0.1
chromadb==0.5.0
pytz==2024.1
pydantic[email]==2.7.1
cryptography==42.0.7
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0
hypothesis==6.103.1
pytest==8.2.1
pytest-asyncio==0.23.7
python-dotenv==1.0.1
email-validator==2.1.1
